requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
sqlalchemy==2.0.17
pydantic==1.10.4
python-dotenv==1.0.0
pytest==7.2.2
aiohttp==3.8.1
starlette==0.26.1
databases==0.6.2
